Struct http_body::Limited

pub struct Limited<B> { /* private fields */ }

A length limited body.

This body will return an error if more than the configured number of bytes are returned on polling the wrapped body.

Implementations

impl<B> Limited<B>

pub fn new(inner: B, limit: usize) -> Self

Create a new Limited.