Crate http_body

Asynchronous HTTP request or response body.

See Body for more details.

Modules

combinators

Combinators for the Body trait.

Structs

Collected

A collected body produced by Body::collect which collects all the DATA frames and trailers.

Data

Future that resolves to the next data chunk from Body

Empty

A body that is always empty.

Full

A body that consists of a single chunk.

LengthLimitError

An error returned when body length exceeds the configured limit.

Limited

A length limited body.

SizeHint

A Body size hint

Trailers

Future that resolves to the optional trailers from Body

Traits

Body

Trait representing a streaming body of a Request or Response.