Trait http_body::Body

pub trait Body {
    type Data: Buf;
    type Error;
    fn poll_data(
        self: Pin<&mut Self>, 
        cx: &mut Context
    ) -> Poll<Option<Result<Self::Data, Self::Error>>>;

    fn poll_trailers(
        self: Pin<&mut Self>, 
        cx: &mut Context
    ) -> Poll<Result<Option<HeaderMap>, Self::Error>>;

    fn is_end_stream(&self) -> bool { ... }

    fn size_hint(&self) -> SizeHint { ... }

    fn data(&mut self) -> Data<Self>
    where
        Self: Unpin + Sized,
    { ... }

    fn trailers(&mut self) -> Trailers<Self>
    where
        Self: Unpin + Sized,
    { ... }
}

Trait representing a streaming body of a Request or Response.

Data is streamed via the poll_data function, which asynchronously yields T: Buf values. The size_hint function provides insight into the total number of bytes that will be streamed.

The poll_trailers function returns an optional set of trailers used to finalize the request / response exchange. This is mostly used when using the HTTP/2.0 protocol.

Associated Types

type Data: Buf

Values yielded by the Body.

type Error

The error type this Body might generate.

Loading content...
Required methods

fn poll_data(
self: Pin<&mut Self>,
cx: &mut Context
) -> Poll<Option<Result<Self::Data, Self::Error>>>

Attempt to pull out the next data buffer of this stream.

fn poll_trailers(
self: Pin<&mut Self>,
cx: &mut Context
) -> Poll<Result<Option<HeaderMap>, Self::Error>>

Poll for an optional single HeaderMap of trailers.

This function should only be called once poll_data returns None.

Loading content...
Provided methods

fn is_end_stream(&self) -> bool

Returns true when the end of stream has been reached.

An end of stream means that both poll_data and poll_trailers will return None.

A return value of false does not guarantee that a value will be returned from poll_stream or poll_trailers.

fn size_hint(&self) -> SizeHint

Returns the bounds on the remaining length of the stream.

When the exact remaining length of the stream is known, the upper bound will be set and will equal the lower bound.

fn data(&mut self) -> Data<Self> where
Self: Unpin + Sized,

Returns future that resolves to next data chunk, if any.

fn trailers(&mut self) -> Trailers<Self> where
Self: Unpin + Sized,

Returns future that resolves to trailers, if any.

Loading content...