Trait http_body::Body

pub trait Body {
    type Data: Buf;
    type Error;

    // Required methods
    fn poll_data(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>
    ) -> Poll<Option<Result<Self::Data, Self::Error>>>;
    fn poll_trailers(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>
    ) -> Poll<Result<Option<HeaderMap>, Self::Error>>;

    // Provided methods
    fn is_end_stream(&self) -> bool { ... }
    fn size_hint(&self) -> SizeHint { ... }
    fn data(&mut self) -> Data<'_, Self> 
       where Self: Unpin + Sized { ... }
    fn trailers(&mut self) -> Trailers<'_, Self> 
       where Self: Unpin + Sized { ... }
    fn map_data<F, B>(self, f: F) -> MapData<Self, F>
       where Self: Sized,
             F: FnMut(Self::Data) -> B,
             B: Buf { ... }
    fn map_err<F, E>(self, f: F) -> MapErr<Self, F>
       where Self: Sized,
             F: FnMut(Self::Error) -> E { ... }
    fn collect(self) -> Collect<Self>
       where Self: Sized { ... }
    fn boxed(self) -> BoxBody<Self::Data, Self::Error>
       where Self: Sized + Send + Sync + 'static { ... }
    fn boxed_unsync(self) -> UnsyncBoxBody<Self::Data, Self::Error>
       where Self: Sized + Send + 'static { ... }
}

Trait representing a streaming body of a Request or Response.

Data is streamed via the poll_data function, which asynchronously yields T: Buf values. The size_hint function provides insight into the total number of bytes that will be streamed.

The poll_trailers function returns an optional set of trailers used to finalize the request / response exchange. This is mostly used when using the HTTP/2.0 protocol.

Required Associated Types

type Data: Buf

Values yielded by the Body.

type Error

The error type this Body might generate.

Required Methods

fn poll_data( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> Poll<Option<Result<Self::Data, Self::Error>>>

Attempt to pull out the next data buffer of this stream.

fn poll_trailers( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> Poll<Result<Option<HeaderMap>, Self::Error>>

Poll for an optional single HeaderMap of trailers.

This function should only be called once poll_data returns None.

Provided Methods

fn is_end_stream(&self) -> bool

Returns true when the end of stream has been reached.

An end of stream means that both poll_data and poll_trailers will return None.

A return value of false does not guarantee that a value will be returned from poll_stream or poll_trailers.

fn size_hint(&self) -> SizeHint

Returns the bounds on the remaining length of the stream.

When the exact remaining length of the stream is known, the upper bound will be set and will equal the lower bound.

fn data(&mut self) -> Data<'_, Self>
where Self: Unpin + Sized,

Returns future that resolves to next data chunk, if any.

fn trailers(&mut self) -> Trailers<'_, Self>
where Self: Unpin + Sized,

Returns future that resolves to trailers, if any.

fn map_data<F, B>(self, f: F) -> MapData<Self, F>
where Self: Sized, F: FnMut(Self::Data) -> B, B: Buf,

Maps this body’s data value to a different value.

fn map_err<F, E>(self, f: F) -> MapErr<Self, F>
where Self: Sized, F: FnMut(Self::Error) -> E,

Maps this body’s error value to a different value.

fn collect(self) -> Collect<Self>
where Self: Sized,

Turn this body into Collected body which will collect all the DATA frames and trailers.

fn boxed(self) -> BoxBody<Self::Data, Self::Error>
where Self: Sized + Send + Sync + 'static,

Turn this body into a boxed trait object.

fn boxed_unsync(self) -> UnsyncBoxBody<Self::Data, Self::Error>
where Self: Sized + Send + 'static,

Turn this body into a boxed trait object that is !Sync.